refactor: Simplify ProjectRegister component by extracting form logic into a separate component
- Removed inline form logic and state management from ProjectRegister.
- Introduced ProjectRegisterForm component to encapsulate form functionality and improve code organization.
- Streamlined imports and reduced component size for better maintainability.
